Simulator class error fix
